[Optim] Add an option controlling the optimizer eps (#917)
### What does this PR do?

Add an option into `torchtitan/components/optimizer.py` and config files
to allow users to control the epsilon value passed into the optimizer.


### Why is this change necessary?
Some papers suggest that providing users with the flexibility to control
the epsilon value is important. This ensures that epsilon does not
dominate the updates when gradients become extremely small. The default
value of $1^{-8}$ may not always be sufficiently small; in some cases,
setting epsilon to $1^{-15}$ or even zero could be more appropriate.


![image](https://github.com/user-attachments/assets/25bfaf27-3006-45a0-b63a-cbf9216f20ef)

See https://arxiv.org/pdf/2412.17743 and
https://arxiv.org/abs/2309.14322 for more details.

Co-authored-by: Chien-Chin Huang <fegins@gmail.com>